Renamed from shared to dataviews for consistency
